Updated insomnia requests tests, fixed seism model and resource

+++ Seismology/api/main/resources/seism.py
@@ -2,14 +2,17 @@ from flask_restful import Resource
 from flask import request, jsonify
 from .. import db
 from main.models import SeismModel
+from datetime import datetime
 
 
 class VerifiedSeism(Resource):
 
     def get(self, id):
         seism = db.session.query(SeismModel).get_or_404(id)
-        return seism.to_json()
-
+        if seism.verified:
+            return seism.to_json()
+        else:
+            return 'Forbidden access', 403
 
 class VerifiedSeisms(Resource):
 
@@ -22,25 +25,37 @@ class UnverifiedSeism(Resource):
 
     def get(self, id):
         seism = db.session.query(SeismModel).get_or_404(id)
-        return seism.to_json()
+        if not seism.verified:
+            return seism.to_json()
+        else:
+            return 'Forbidden access', 403
 
     def delete(self, id):
         seism = db.session.query(SeismModel).get_or_404(id)
-        db.session.delete(seism)
-        db.session.commit()
-        return "Unverified seism deleted succesfully", 204
+        if not seism.verified:
+            db.session.delete(seism)
+            db.session.commit()
+            return "Unverified seism deleted succesfully", 204
+        else:
+            return 'Forbidden access', 403
 
     def put(self, id):
         seism = db.session.query(SeismModel).get_or_404(id)
-        for key, value in request.get_json().items():
-            setattr(seism, key, value)
-        db.session.add(seism)
-        db.session.commit()
-        return seism.to_json(), 201
+        if not seism.verified:
+            for key, value in request.get_json().items():
+                if key == 'datetime':
+                    setattr(seism, key, datetime.strptime(value, "%Y-%m-%d %H:%M:%S"))
+                else:
+                    setattr(seism, key, value)
+            db.session.add(seism)
+            db.session.commit()
+            return seism.to_json(), 201
+        else:
+            return 'Forbidden access', 403
 
 
 class UnverifiedSeisms(Resource):
 
     def get(self):
         seisms = db.session.query(SeismModel).filter(SeismModel.verified == False).all()
-        return jsonify({'Unverified-Seisms': [seism.to_json for seism in seisms]})
+        return jsonify({'Unverified-Seisms': [seism.to_json() for seism in seisms]})
